+++ set/set.h
@@ -12,7 +12,7 @@ struct set set__empty(void);
 int set__is_empty( const struct set* set);
 
 int set__find( const struct set* set, int e);
-size_t set__size( const struct set* set);
+int set__size( const struct set* set);
 
 int set__add( struct set* set, int e);
 int set__remove( struct set* set, int e);
